feat(napi/transform): align output `SourceMap` with Rollup's `ExistingRawSourceMap` (#5657)
closes #5578
