+++ napi/transform/index.d.ts
@@ -136,10 +136,12 @@ export interface ReactRefreshBindingOptions {
 export interface SourceMap {
   file?: string
   mappings?: string
+  names?: Array<string>
   sourceRoot?: string
   sources?: Array<string | undefined | null>
   sourcesContent?: Array<string | undefined | null>
-  names?: Array<string>
+  version: number
+  x_google_ignoreList?: Array<number>
 }
 
 /**
